# docs/swagger.yaml
definitions:
...
  oauth.TrustLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-varnames:
    - TrustLevelNewUser
    - TrustLevelBasicUser
    - TrustLevelUser
    - TrustLevelActiveUser
    - TrustLevelLeader
...
  project.CreateProjectRequestBody:
    properties:
      allow_same_ip:
        type: boolean
      description:
        maxLength: 1024
        type: string
      distribution_type:
        allOf:
        - $ref: '#/definitions/project.DistributionType'
        enum:
        - 0
        - 1
      end_time:
        type: string
      hide_from_explore:
        type: boolean
      minimum_trust_level:
        allOf:
        - $ref: '#/definitions/oauth.TrustLevel'
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
      name:
        maxLength: 32
        minLength: 1
        type: string
      project_items:
        items:
          type: string
        minItems: 1
        type: array
      project_tags:
        items:
          type: string
        type: array
      risk_level:
        maximum: 100
        minimum: 0
        type: integer
      start_time:
        type: string
      topic_id:
        type: integer
    required:
    - end_time
    - name
    - project_items
    - start_time
    type: object
  project.DistributionType:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - DistributionTypeOneForEach
    - DistributionTypeLottery
    - DistributionTypeInvite
  project.GetProjectResponseData:
    properties:
      allow_same_ip:
        type: boolean
      available_items_count:
        type: integer
      created_at:
        type: string
      creator_id:
        type: integer
      creator_nickname:
        type: string
      creator_username:
        type: string
      description:
        type: string
      distribution_type:
        $ref: '#/definitions/project.DistributionType'
      end_time:
        type: string
      hide_from_explore:
        type: boolean
      id:
        type: string
      is_completed:
        type: boolean
      is_received:
        type: boolean
      minimum_trust_level:
        $ref: '#/definitions/oauth.TrustLevel'
      name:
        type: string
      received_content:
        type: string
      report_count:
        type: integer
      risk_level:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/project.ProjectStatus'
      tags:
        items:
          type: string
        type: array
      total_items:
        type: integer
      updated_at:
        type: string
    type: object
  project.ListProjectsResponse:
    properties:
      data:
        $ref: '#/definitions/project.ListProjectsResponseData'
      error_msg:
        type: string
    type: object
  project.ListProjectsResponseData:
    properties:
      results:
        items:
          $ref: '#/definitions/project.ListProjectsResponseDataResult'
        type: array
      total:
        type: integer
    type: object
  project.ListProjectsResponseDataResult:
    properties:
      allow_same_ip:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      distribution_type:
        $ref: '#/definitions/project.DistributionType'
      end_time:
        type: string
      hide_from_explore:
        type: boolean
      id:
        type: string
      minimum_trust_level:
        $ref: '#/definitions/oauth.TrustLevel'
      name:
        type: string
      risk_level:
        type: integer
      start_time:
        type: string
      tags:
        items:
          type: string
        type: array
      total_items:
        type: integer
    type: object
  project.ListReceiveHistoryResponse:
    properties:
      data:
        $ref: '#/definitions/project.ListReceiveHistoryResponseData'
      error_msg:
        type: string
    type: object
  project.ListReceiveHistoryResponseData:
    properties:
      results:
        items:
          $ref: '#/definitions/project.ListReceiveHistoryResponseDataResult'
        type: array
      total:
        type: integer
    type: object
  project.ListReceiveHistoryResponseDataResult:
    properties:
      content:
        type: string
      project_creator:
        type: string
      project_creator_nickname:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      received_at:
        type: string
    type: object
...
  project.ProjectResponse:
    properties:
      data: {}
      error_msg:
        type: string
    type: object
  project.ProjectStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - ProjectStatusNormal
    - ProjectStatusHidden
    - ProjectStatusViolation
  project.ReportProjectRequestBody:
    properties:
      reason:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - reason
    type: object
  project.UpdateProjectRequestBody:
    properties:
      allow_same_ip:
        type: boolean
      description:
        maxLength: 1024
        type: string
      enable_filter:
        type: boolean
      end_time:
        type: string
      hide_from_explore:
        type: boolean
      minimum_trust_level:
        allOf:
        - $ref: '#/definitions/oauth.TrustLevel'
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
      name:
        maxLength: 32
        minLength: 1
        type: string
      project_items:
        items:
          type: string
        type: array
      project_tags:
        items:
          type: string
        type: array
      risk_level:
        maximum: 100
        minimum: 0
        type: integer
      start_time:
        type: string
    required:
    - end_time
    - name
    - start_time
    type: object
info:
  contact: {}
  title: LINUX DO CDK
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListProjectsResponse'
      tags:
      - project
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
    get:
      description: 获取指定项目所有信息以及领取情况 (Get all information and claim status for a specific
        project)
      parameters:
      - description: 项目ID (Project ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/project.ProjectResponse'
            - properties:
                data:
                  $ref: '#/definitions/project.GetProjectResponseData'
              type: object
      summary: 获取指定项目信息 (Get specific project information)
      tags:
      - project
    put:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: string
      - description: 项目信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/project.UpdateProjectRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
  /api/v1/projects/{id}/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
  /api/v1/projects/{id}/receivers:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
  /api/v1/projects/{id}/report:
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: string
      - description: 举报信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/project.ReportProjectRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
      tags:
      - project
  /api/v1/projects/mine:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListProjectsResponse'
      tags:
      - project
  /api/v1/projects/received:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListReceiveHistoryResponse'
      tags:
      - project
...
swagger: "2.0"